pyproject.toml
[project]
name = "wows-discordbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.5.2",
    "parse>=1.20.2",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "debugpy>=1.8.13",
    "isort>=6.0.1",
]
